kill only temp buffers in vdiff-magit-compare
Instead of just killing both buffers when vdiff-buffers
finishes (KILL-BUFFERS-ON-QUIT=t), preserve any buffers that are
pointing to files.

This is predicated on the assumption that most times, e.g. `M-x
vdiff-magit-compare` is running against an open file, comparing it to
a revision, and then that the user would like to preserve the original
file buffer they were using before starting vdiff.